Dev task 0009 others

+++ app/Http/Controllers/Admin/Account/Admin/AdminController.php
@@ -27,7 +27,7 @@ class AdminController extends Controller
     public function index(): Response
     {
         $admins = Admin::select('id', 'name', 'kana', 'email')
-            ->orderBy('kana')
+            ->orderBy('id')
             ->get();
 
         return Inertia::render('Admin/Account/Admin/Index', [
@@ -58,6 +58,8 @@ class AdminController extends Controller
                     'kana' => $request->kana,
                     'email' => $request->email,
                     'password' => Hash::make($password),
+                    'created_by' => Auth::guard('admins')->user()->name,
+                    'updated_by' => Auth::guard('admins')->user()->name,
                 ]);
             });
 
@@ -88,6 +90,10 @@ class AdminController extends Controller
                 'name' => $admin->name,
                 'kana' => $admin->kana,
                 'email' => $admin->email,
+                'created_at' => $admin->created_at->format('Y-m-d H:i:s'),
+                'created_by' => $admin->created_by,
+                'updated_at'=> $admin->updated_at->format('Y-m-d H:i:s'),
+                'updated_by' => $admin->updated_by,
             ],
         ]);
     }
@@ -111,16 +117,27 @@ class AdminController extends Controller
     /**
      * Update the specified resource in storage.
      */
-    public function update(UpdateAdminRequest $request, Admin $admin): RedirectResponse
+    public function update(UpdateAdminRequest $request, String $id): RedirectResponse
     {
+        $admin = Admin::find($id);
+
+        if (!$admin) {
+            return to_route('admin.account.admins.index')->with([
+                'flash_id' => Str::uuid(),
+                'flash_message' => '対象のデータが見つかりません',
+                'flash_status' => 'error',
+            ]);
+        }
+
         try {
             DB::transaction(function () use ($request, $admin) {
-                if ($admin->updated_at->format('Y-m-d H:i:s') !== $request->updatedAt) {
+                if ($admin->updated_at->format('Y-m-d H:i:s') !== $request->updated_at) {
                     throw new OptimisticLockException;
                 }
 
                 $admin->name = $request->name;
                 $admin->kana = $request->kana;
+                $admin->updated_by = Auth::guard('admins')->user()->name;
 
                 $admin->save();
             });
@@ -148,8 +165,18 @@ class AdminController extends Controller
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(Admin $admin): RedirectResponse
+    public function destroy(String $id): RedirectResponse
     {
+        $admin = Admin::find($id);
+
+        if (!$admin) {
+            return to_route('admin.account.admins.index')->with([
+                'flash_id' => Str::uuid(),
+                'flash_message' => '対象のデータが見つかりません',
+                'flash_status' => 'error',
+            ]);
+        }
+
         try {
             DB::transaction(function () use ($admin) {
                 $admin->delete();
@@ -178,7 +205,7 @@ class AdminController extends Controller
             'ids.required' => '削除対象を選択してください',
             'ids.array' => '削除対象の形式が正しくありません',
             'ids.*.integer' => '無効なIDが含まれています',
-            'ids.*.exists' => '存在しないデータが選択されています。ページを更新して再試行してください。',
+            'ids.*.exists' => '存在しないデータが選択されていたのでページを更新しました。再試行してください。',
         ]);
 
         try {
