+++ app/Http/Controllers/Admin/Office/OfficeController.php
@@ -25,7 +25,7 @@ class OfficeController extends Controller
     public function index(): Response
     {
         $offices = Office::select('id', 'name', 'kana')
-            ->orderBy('kana')
+            ->orderBy('id')
             ->get();
 
         return Inertia::render('Admin/Office/Index', [
@@ -51,6 +51,8 @@ class OfficeController extends Controller
                 Office::create([
                     'name' => $request->name,
                     'kana' => $request->kana,
+                    'created_by' => Auth::guard('admins')->user()->name,
+                    'updated_by' => Auth::guard('admins')->user()->name,
                 ]);
             });
 
@@ -78,6 +80,10 @@ class OfficeController extends Controller
                 'id' => $office->id,
                 'name' => $office->name,
                 'kana' => $office->kana,
+                'created_at' => $office->created_at->format('Y-m-d H:i:s'),
+                'created_by' => $office->created_by,
+                'updated_at'=> $office->updated_at->format('Y-m-d H:i:s'),
+                'updated_by' => $office->updated_by,
             ],
         ]);
     }
@@ -100,16 +106,27 @@ class OfficeController extends Controller
     /**
      * Update the specified resource in storage.
      */
-    public function update(UpdateOfficeRequest $request, Office $office): RedirectResponse
+    public function update(UpdateOfficeRequest $request, String $id): RedirectResponse
     {
+        $office = Office::find($id);
+
+        if (!$office) {
+            return to_route('admin.offices.index')->with([
+                'flash_id' => Str::uuid(),
+                'flash_message' => '対象のデータが見つかりません',
+                'flash_status' => 'error',
+            ]);
+        }
+
         try {
             DB::transaction(function () use ($request, $office) {
-                if ($office->updated_at->format('Y-m-d H:i:s') !== $request->updatedAt) {
+                if ($office->updated_at->format('Y-m-d H:i:s') !== $request->updated_at) {
                     throw new OptimisticLockException;
                 }
 
                 $office->name = $request->name;
                 $office->kana = $request->kana;
+                $office->updated_by = Auth::guard('admins')->user()->name;
 
                 $office->save();
             });
@@ -137,8 +154,18 @@ class OfficeController extends Controller
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(Office $office): RedirectResponse
+    public function destroy(String $id): RedirectResponse
     {
+        $office = Office::find($id);
+
+        if (!$office) {
+            return to_route('admin.offices.index')->with([
+                'flash_id' => Str::uuid(),
+                'flash_message' => '対象のデータが見つかりません',
+                'flash_status' => 'error',
+            ]);
+        }
+
         try {
             DB::transaction(function () use ($office) {
                 $office->delete();
@@ -167,7 +194,7 @@ class OfficeController extends Controller
             'ids.required' => '削除対象を選択してください',
             'ids.array' => '削除対象の形式が正しくありません',
             'ids.*.integer' => '無効なIDが含まれています',
-            'ids.*.exists' => '存在しないデータが選択されています。ページを更新して再試行してください。',
+            'ids.*.exists' => '存在しないデータが選択されていたのでページを更新しました。再試行してください。',
         ]);
 
         try {
